workflows/dagster/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dagster-pipeline"
version = "0.1.0"
description = "Data orchestration pipeline with Dagster for ETL, maintenance, and snapshots"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Rodgers Nyangweso", email = "team@example.com"}
]
classifiers = [
    "Programming Language :: Python :: 3.10"
]

# Core dependencies (production)
dependencies = [
    # Core Dagster
    "dagster==1.12.4",
    "dagster-webserver==1.12.4",
    "dagster-postgres==0.28.4",
    
    # AWS/ECS Support (for production deployment)
    "dagster-aws==0.28.4",      # ECS run launcher, S3 compute logs
    "boto3==1.35.36",           # AWS SDK
    
    # Database drivers
    "psycopg2-binary==2.9.10",
    "pymysql==1.1.1",
    "clickhouse-driver==0.2.9",
    "clickhouse-connect==0.8.18",
    "mysql-connector-python==8.0.33",
    
    # Data processing
    "pyarrow==17.0.0",
    
    # Utilities
    "pyyaml==6.0.2",
    "cryptography>=41.0.0",     # Required for PyMySQL caching_sha2_password auth
    "python-dotenv==1.0.1",
    "pydantic==2.9.2",
    "sqlalchemy==2.0.43",
    "tenacity==9.1.2",
    
    # Monitoring
    "structlog==25.5.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest==8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black==25.11.0",
    "ruff==0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

# All optional dependencies
all = [
    "dagster-pipeline[dev]",
]

[project.urls]
Homepage = "https://github.com/yourorg/dagster-pipeline"
Repository = "https://github.com/yourorg/dagster-pipeline"
Documentation = "https://github.com/yourorg/dagster-pipeline/tree/main/docs"

# ============================================================================
# Tool Configurations
# ============================================================================

[tool.setuptools.packages.find]
where = ["."]
include = ["dagster_pipeline*"]
exclude = ["tests*", "docs*"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.ruff.lint.isort]
known-first-party = ["dagster_pipeline"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=dagster_pipeline",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start permissive, tighten over time
ignore_missing_imports = true
files = ["dagster_pipeline"]

[tool.coverage.run]
source = ["dagster_pipeline"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]